get_cpu_count: quiet -Wshorten-64-to-32 warnings
sysconf returns a long; cast (unsigned) dwNumberOfProcessors to int for
good measure

+++ vp8/common/generic/systemdependent.c
@@ -37,9 +37,9 @@ static int get_cpu_count() {
 
 #if HAVE_UNISTD_H && !defined(__OS2__)
 #if defined(_SC_NPROCESSORS_ONLN)
-  core_count = sysconf(_SC_NPROCESSORS_ONLN);
+  core_count = (int)sysconf(_SC_NPROCESSORS_ONLN);
 #elif defined(_SC_NPROC_ONLN)
-  core_count = sysconf(_SC_NPROC_ONLN);
+  core_count = (int)sysconf(_SC_NPROC_ONLN);
 #endif
 #elif defined(_WIN32)
   {
@@ -61,7 +61,7 @@ static int get_cpu_count() {
       GetSystemInfo(&sysinfo);
 #endif
 
-    core_count = sysinfo.dwNumberOfProcessors;
+    core_count = (int)sysinfo.dwNumberOfProcessors;
   }
 #elif defined(__OS2__)
   {
